Converted the setting into an editor option named dynamicWrapping.
I converted the setting I introduced into a separate config option so
that users can configure it independently of other options. The setting
works in a backwards compatible way so that keeping it off doesn't
change the user experience than what the user is experienced to. Only
when the setting is turned on will it have an effect on the experience.