Remove .pp from Ruby extension list. Issue #5366
The *.pp file extension as it pertains it Ruby is actually
a file type used to denote a Puppet Manifest written in the Puppet
DSL (which itself is written in Ruby). The affected setting was
overriding VSCode Puppet extensions that are expecting the .pp
extension to be detected as file type Puppet.