Use common TS install for both the TS Extension and Html Extension (#21578)
* Use common TS install for both the TS Extension and Html Extension

**Bug**
At least two version of Typescript are shipped in our extensions: one in the typescript extension and one for the html extension. This adds about 5MB to package install size and also results in inconsistent behavior

**Fix**
Change the TypeScript extension to also use the common version of TypeScript. Bump this version up to 2.2.1

* Don't hardcode paths for typescript