prevent reload window closes the debuggee; fixes #23390
I cannot reproduce the problem when running out of source, so I could not really verify that the fix works.

`environmentService.debugExtensionHost.break` is used for two things: 1) as an indication that debug mode is active and 2) it indicates that break mode is active. The first use is now broken, because debug mode can be enabled with the command line flag `--debugPluginHost` too.  This dual use forbids that we can just add a test for `--debugPluginHost` when parsing the value from the command line. For this reason I'm following @bpasero's suggestion to use the `debugExtensionHost.port` as an indication that debug mode is active.