Clean Up JSDoc Completion Provider Implementation (#25981)
* Fix JSDoc Completion Provider

**Bug**
ae6151aef448a99c94dd06dddf984e4734e9609e fixed #11944 by changing how enter works in completion providers. Now if the completion exactly matches the current text, pressing `enter` is not captured by the the suggestion widget. Instead it inserts a new line into the editor. This broke the jsdoc completion provider which uses `''` as a placeholder insertion before calculating the actual insertion upon being accepted. This weird behavior is because of a of the current tsserver api

**Fix**
Change the insert text to a single space instead. This seems to fix this issue in my testing.

* Slightly more elegant fix