Update paths.js to allow use of ENV variable instead of command-line override for userDataPath
This patch allows defining the environment variable "vscode_config_prefix" as a global default, instead of relying on command line arguments.  The solution is similar to NPM's npm_config_prefix environment variable.  if defined, APPDATA, XDG_CONFIG_HOME, and other solutions are ignored for determining the path value.  This is 1/2 the solution to removing the user data from the Windows roaming profile.