Store expansion state globally only when there are items.
When a debugger needs more time while stepping, the tree will go empty first
(while running) and then come back with data (when stopped again). This
modifies the change for #16031 to allow the tree to properly restore
state in case a debugger needs more time providing content.

The expanded state is saved in a global variable. A check ensures that an
intermittent empty tree does not override the global variable. Future refreshes
of an empty tree will then restore the expanded state of variables.