Adding missed dependency libxss1
In releasing for Chromebook I'm working on a minimal installation, so the base distribution has less packages pre-installed.

libxss1 is both a compile-time and run-time dependency, which would appear to have been missed (on my system it causes run-time errors if it's not present, due to the absence of libXss.so.1). I would expect this to cause the same issue on other minimal Debian installations.