Switch VS Code workspace to use typescript-tslint-plugin
Use the plugin version of TS lint rather than the external version. Add the plugin as a local dev dependency and add the vscode extension that contributes the plugin as a recomended extension. This should enable the plugin when using the standard version of TS and when using the workspace version of TS