editor - ensure contract of openEditor()/openEditors() is respected
These methods return an IEditor as a result but even if the opening fails or the editor was not opened as active.

The fix is to check for the result of the open call and only return the editor if the editor was opened actually.

This restores the behaviour we had before GRID.

fixes #64767