Make sure `Dispose.register` does not allow registering new disposables after the base has been disposed
`register` can be called after `dispose` in some race conditions, especially when dealing with events. Previously in this case, `register` would continue to hold onto the newly registered objects and they would never be properly disposed. With this change, we now dispose of newly registered objects immediately and print a warning if the base has already been disposed