Use ReadonlyArrays for IActions used in notifications
These arrays of actions should not be modified by the caller. Mark them as readonly to prevent this. This also let me eliminate a number of not null supressions that had been introduced during the strict null migration

One small code change was required in `showError`. However I believe this was actually a bug since modifying the input array was likely not what would be expected (in practice, we usually just passed in array literals so I don't think we hit this one)