Fix quick fix problems menu button when using custom context menus
Fixes #77018

**Bug**
Code actions now have a life cycle. As part of this change, we had to add lifecycle tracking to the quick fix button in the problems hover view.  This works by disposing of the code actions when the hover is closed

When using native context menus, clicking the quick fix button in a problem hover keeps the hover open. However when using custom context menus, the hover is closed. This means that the list of code actions we are now displaying has been disposed of and will not function as expected

**Fix**
Refactor to route the UI for showing the code actions menu through the code actions controller, which properly handles the lifecycle of code actions. This also somewhat reduces code duplication