debug - fix active line for column-free breakpoints not be decorated correctly
In #81718 we adjusted breakpoint display logic to show an 'empty' arrow on the
line, and the arrow with a dot at the column where the debugger is running.
However, not all debuggers support column breakpoints. Bowden noticed, while
working on his Python demo for ignite, that breakpoints in Python code no
longer displayed correctly.

This PR tweaks the logic so that if we see a breakpoint that doesn't have
an associated column, we show the arrow with a dot on the line, rather than
assuming that there'll be an inline indicator arrow.