files - speed up reads and writes
For reading, introduce the readFileStream capability that reduces
the overhead of low level file reading by switching to an event based
model.

For writing, try to consume up to N chunks of data from the underlying
source and directly call writeFile if the source is consumed. This will avoid
the overhead of low level file reading for small to medium files.