Allow for rulers to be individually colored
This change set:
adds an interface for Ruler Color Options,
adds a union type of that interface and number,
adjust the rulers type hint to use said union type,
adds a mechanism to apply boxshadow to fastDomNode, and
updates the way rulers are rendered to support individual colors.

I've also ensured that descriptions are provided for settings.json.
So now, all nodes will provide explanations.