Fix threads not fetched after restarting debug adapter
A RunOnceScheduler (this.fetchThreadsScheduler) is used to throttle
fetching threads from debug adapters. this.fetchThreadsScheduler is
disposed when a debug adapter is restarted because of:

    this.rawListeners.push(this.fetchThreadsScheduler);

This causes threads to never be fetched after restarting the debug
adapter.

This is fixed by setting this.fetchThreadsScheduler to undefined after
it is disposed. This way, a new RunOnceScheduler is created after
restarting the debug adapter.

Fixes #88784