Make sure we dispose of every custom editors input individually
Fixes #88229

There should a 1:1 mapping between VS Code's custom editors and custom editors (specifically the webviews) that extensions know about. This was previously not true as a single `CustomEditorInput` could end up being shared between multiple editor groups.

With this change, we always create a copy of the custom editor input during split, and then add a listener for `onDidCloseEditor` so that we can make sure we get rid of the unique inputs (and notify extensions of this)